Added CloudSession to wrap Session + cloud setup
CloudSession aggregates Session and owns both the Session and the CloudEventListener. CloudEventListener will receive messages and either push them into the Session as events, or call waiting listeners with MQTT responses to REST calls.

At this point some of the setup is done, but no calls through to the scanner yet.