Implements cloud block downloading
When scanning from a local scanner, blocks are retrieved using the readImageBlock command, which returns the image info and the image in one MIME multipart response.

With a cloud scanner, the readImageBlock command returns just the JSON part of the response, and includes an imageBlockId. This is used to request the actual image data from the cloud service.

Refactored BlockDownloader to share most of the common implementation, and just have different methods of getting the image data.